+++ src/web/assets/cp/src/js/BaseElementIndex.js
@@ -56,7 +56,7 @@ Craft.BaseElementIndex = Garnish.Base.extend(
     $scoreSortAttribute: null,
     $structureSortAttribute: null,
 
-    _sourcePath: null,
+    sourcePaths: null,
     $sourcePathOuterContainer: null,
     $sourcePathInnerContainer: null,
     $sourcePathOverflowBtnContainer: null,
@@ -97,6 +97,8 @@ Craft.BaseElementIndex = Garnish.Base.extend(
       this.elementType = elementType;
       this.$container = $container;
       this.setSettings(settings, Craft.BaseElementIndex.defaults);
+
+      this.sourcePaths = {};
 
       // Set the state objects
       // ---------------------------------------------------------------------
@@ -301,15 +303,68 @@ Craft.BaseElementIndex = Garnish.Base.extend(
       // Select the initial source + source path
       // ---------------------------------------------------------------------
 
+      // Grab the localStorage step key up front, so we don's lose track of it when the default source's default
+      // source path is selected
+      const stepKey = this.getSelectedSourceState('sourcePathStep');
+
       this.selectDefaultSource();
 
       const sourcePath = this.getDefaultSourcePath();
-      if (sourcePath !== null) {
+
+      // If no default source path was explicitly configured, use the localStorage key
+      if (!sourcePath && stepKey) {
+        this.loadSourcePathByKey(stepKey).then((sourcePath) => {
+          if (sourcePath) {
+            // Filter out any source path steps that are above the source's root
+            const lastSourceKey = this.sourceKey.split('/').slice(-1)[0];
+            const sourceRootIndex = sourcePath.findIndex(
+              (p) => p.key === lastSourceKey
+            );
+            if (sourceRootIndex !== -1) {
+              this.sourcePath = sourcePath.slice(sourceRootIndex);
+            }
+          }
+          this.afterSetInitialSource();
+        });
+      } else {
         this.sourcePath = sourcePath;
+        this.afterSetInitialSource();
       }
+    },
+
+    afterInit: function () {
+      this.onAfterInit();
+    },
+
+    loadSourcePathByKey: function (stepKey) {
+      return new Promise((resolve, reject) => {
+        // If the step key is equal to the current source key, then it represents the root. No source path needed.
+        if (stepKey === this.sourceKey) {
+          resolve([]);
+          return;
+        }
+
+        const params = this.getViewParams();
+        params.stepKey = stepKey;
+
+        Craft.sendActionRequest('POST', 'element-indexes/source-path', {
+          data: params,
+        })
+          .then(({data}) => {
+            resolve(data.sourcePath);
+          })
+          .catch(reject);
+      });
+    },
+
+    afterSetInitialSource: function () {
+      // Resize handler
+      // ---------------------------------------------------------------------
+
       if (this.settings.context === 'index') {
         this.addListener(Garnish.$win, 'resize', 'handleResize');
       }
+
       this.handleResize();
 
       // Load the first batch of elements!
@@ -319,10 +374,6 @@ Craft.BaseElementIndex = Garnish.Base.extend(
       this.setPage(Craft.pageNum);
 
       this.updateElements(true);
-    },
-
-    afterInit: function () {
-      this.onAfterInit();
     },
 
     handleResize: function () {
@@ -485,6 +536,16 @@ Craft.BaseElementIndex = Garnish.Base.extend(
     },
 
     getDefaultSourceKey: function () {
+      if (
+        this.settings.preferStoredSource &&
+        this.instanceState.selectedSource
+      ) {
+        // Discard the defaults and go with localStorage
+        this.settings.defaultSource = null;
+        this.settings.defaultSourcePath = null;
+        return this.instanceState.selectedSource;
+      }
+
       let sourceKey = null;
 
       if (this.settings.defaultSource) {
@@ -553,14 +614,15 @@ Craft.BaseElementIndex = Garnish.Base.extend(
      * @returns {Object[]}
      */
     get sourcePath() {
-      return this._sourcePath || [];
+      return this.sourcePaths[this.sourceKey] || [];
     },
 
     /**
      * @param {Object[]|null} sourcePath
      */
     set sourcePath(sourcePath) {
-      this._sourcePath = sourcePath && sourcePath.length ? sourcePath : null;
+      this.sourcePaths[this.sourceKey] =
+        sourcePath && sourcePath.length ? sourcePath : null;
 
       if (this.$sourcePathOuterContainer) {
         this.$sourcePathOuterContainer.remove();
@@ -570,7 +632,7 @@ Craft.BaseElementIndex = Garnish.Base.extend(
         this.$sourcePathActionsBtn = null;
       }
 
-      if (this._sourcePath && this.settings.showSourcePath) {
+      if (this.sourcePaths[this.sourceKey] && this.settings.showSourcePath) {
         const actions = this.getSourcePathActions();
 
         this.$sourcePathOuterContainer = $('<div/>', {
@@ -750,6 +812,16 @@ Craft.BaseElementIndex = Garnish.Base.extend(
           );
         }
       }
+
+      // Store the source path
+      this.setSelecetedSourceState(
+        'sourcePathStep',
+        (this.sourcePaths[this.sourceKey]
+          ? this.sourcePaths[this.sourceKey][
+              this.sourcePaths[this.sourceKey].length - 1
+            ].key
+          : null) || null
+      );
 
       this.onSourcePathChange();
     },
@@ -1601,7 +1673,9 @@ Craft.BaseElementIndex = Garnish.Base.extend(
 
       this.selectViewMode(viewMode);
 
-      this.sourcePath = this.$source.data('default-source-path');
+      this.sourcePath =
+        this.sourcePaths[this.sourceKey] ||
+        this.$source.data('default-source-path');
 
       this.onSelectSource();
 
@@ -2755,6 +2829,7 @@ Craft.BaseElementIndex = Garnish.Base.extend(
       defaultSiteId: null,
       defaultSource: null,
       defaultSourcePath: null,
+      preferStoredSource: false,
       showSourcePath: true,
       canHaveDrafts: false,